Move Message class to a new file. Add metatext to identify the sender
