+++ lib/widgets/widgets.dart
@@ -7,3 +7,4 @@ export './raid_form.dart';
 export './raid_table.dart';
 export './raid_console.dart';
 export './raid_chat.dart';
+export './message.dart';
